+++ src/modules/auth/index.ts
@@ -1,0 +1,8 @@
+import router from '@/router'
+import ErrorView from './ErrorView.vue'
+
+router.addRoute({
+  path: '/error',
+  name: 'auth:error',
+  component: ErrorView
+})