+++ src/modules/auth/types.ts
@@ -1,5 +1,12 @@
 import { OrganisationRole } from '../organisations/types'
 
+export interface IDjangoMessage {
+  level: number
+  level_tag: 'debug' | 'info' | 'success' | 'warning' | 'error'
+  message: string
+  tags: string
+}
+
 export interface IUser {
   pk: number
   first_name: string